Move decode test to an integration test
